requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-dateutil==2.8.2